iQEcBAABCAAGBQJYhntNAAoJED0WkGtPHFyzJq0IAI4L4B9GntyYe5GMfNLihVsK  H3RENLINUAKvbwfwtkY7iXagCL+c/DnB4Khc+WvFOgfrcnI3Gm6z+aIw1YIh9DnT  I/yyzqEYBs+G1nhSWUt3Ie0lcQVKY88TdGPl/VG4adb8VqsNDMm04VWh/zioYk17  6gYB+QjhG90sXttzes3DFOulGbqeR55xi2kRY8rkFDgnukMw94mwyT0zCQN1EYzb  FfAUxxWQtJbavMhvChjwK3lVCqdOLLK2ac9n4JNkmoaoQXY0+tuCbxnSn73T7EK5  1eFN4MRad+A97jxDDGqfSg0FWyBUNDxR6SngtCnt/qa6lLc6rD+1LoO8DXsB4Xs=  =CD7/  -----END PGP SIGNATURE-----
Change the string serialization of the Nothing selector

The previous option was not serializable, and was used in questionable
scenarios to represent what we also use Everything() for. The individual
context needs to make the decision about everything vs nothing (and
nothing has no real concept in our API today) so this seems safe.