Fixed clock.fakeTimer.Stop and Reset
Fixed the way Stop identified the fakeClockWaiter to remove, and the
logic for the bool to return.

Fixed Reset to side-effect the waiter in the clock's slice rather than
the original copy.

Added a test func for FakeClock timers.

Fixed Reset too

Improved comments in clock.go to address reviewer comments

Hopefully the updated comments make clear the answers to recent
reviewer comments.

Kubernetes-commit: 3e898d69210918433fd0cd0aa89392faac61b83a