Simplify GroupVersion.Seting()
The old code had special cases that returned the same values as the regular case does.

If group is empty, the conditional
```if len(gv.Group) > 0 {```
doesn't apply, and `gv.Version` is returned.

In the empty case, the special case was returning `""`, which is exactly what `gv.Version` contains.
In the legacy v1 case, the special case was returning `"v1"`, which is exactly what `gv.Version` contains.

Kubernetes-commit: 445044db9ce79097f43ab809ae80e11af43a219c