Fix FakeClock::Reset to always succeed
Current FakeClock::Reset only successfully resets the timer and
returns true when the timer has neither fired nore been stopped.

This is incorrect behavior as real clocks allow for reseting
in both of those situations (as long has the channel has been drained).

This is useful in tests that use a fake clock to test
wait.BackofManager, as the timer resets after each iteration of
Backoff() after the timer has already fired.

Kubernetes-commit: 7e7ee9011ac25bc48d19449f69f5f3f94d5772d5