serializer: Recognizer interface should take a buffer
The original intent of the Recognizer interface was to take a stream
and peek into the contents. Over time all of the callers migrated
to instead performing their own peek and then passing the buffer
as a bytes.NewBuffer to the interface as a stream, which was then
implemented as a read on the other side. This resulted in an
excess buffer allocation in general decoder usage.

Update the interface to take a []bytes slice and change all callers
to stop creating buffered readers, then update the JSON serializer
to check the buffer directly for "json-ness".

Kubernetes-commit: 9f288610eba61dd7ea0eb85e6d8b755b2bccf92b