Discard null values in complex objects in strategic patch
In strategic patch, if key does not exist in original, value of this
new key is not distilled by discarding the null values.

That brings about in key creation step, null values are also patched. However,
during the update process of this key in subsequent patches, these null values
are deleted and this hurts the idempotency of strategic patch.

This PR adds discard mechanism for null values if key does not exist
in original. It traverses all nested objects.

Kubernetes-commit: 9b5d9c70fc45c0fe317fdf98fe34422dbc9f2c05