Fix strategic merge patch $deleteFromPrimitiveList bug
If the $deleteFromPrimitiveList directive is used in a strategic
merge patch and the field being deleted from does not have a
patch strategy of merge, strategic merge patch would not perform
the deletion and instead would replace the contents of the field
with the values attempting to be deleted.

This commit changes strategic merge patch to always perform a
deletion when the $deleteFromPrimitiveList directive is used,
regardless of what the patch strategy of the field is.

Kubernetes-commit: 95fe0a158c2ddd614d782967b5d7112222cff648